backend/requirements.txt
openai>=1.0.0
chromadb
pandas
numpy
scikit-learn
pytest
python-dotenv
fastapi
uvicorn
requests
